Changed page folders into files
