+++ frontend/src/images/dashboard_screenshot.png
